Update Product entity inventory type, add data seeding with products, categories, and user roles, add custom exception for inventory validation

+++ src/main/java/dev/abhishek/ecommerce/modules/product/entity/Product.java
@@ -40,7 +40,7 @@ public class Product {
 
     private BigDecimal price;
 
-    private Integer inventory;
+    private Long inventory;
 
     @NotNull
     @Column(length = 1000)
